+++ CHANGELOG.md
@@ -1,4 +1,6 @@
-### **(V.1.3.0 Changes) (1.20.1 Minecraft)**
+### **(V.1.4.0 Changes) (1.20.1 Minecraft)**
 
 #### Items:
-Added entity type tag `tact:primitive_club_stun_immune` that players/packmakers can add entities into so Primitive Club cannot stun them.
+Added CanEatBiomeTreatWhenNotHungry config that is false by default. When on, it will allow Biome Treat to be eaten when full.
+ Ideally, this can be turned on if another mod has overhauled Hunger system in a way that break norma Biome Treat eating.
+ Seriously, this is only a workaround. The Hunger overhauling mods should have made sure their changes did not break other mod's foods.